backend/requirements.txt
# --- Core ---
Django==5.2.* # Dùng bản 5.2.x mới nhất (LTS)
psycopg2-binary            # Driver PostgreSQL
django-unfold              # Modern admin interface

# --- API (Cho Vue.js) ---
djangorestframework>=3.15  # Bản mới hỗ trợ Django 5
django-cors-headers

# --- Auth ---
djangorestframework-simplejwt

# --- Tiện ích ---
python-dotenv
Pillow
django-filter

# --- Editor ---
# Lưu ý: Với Django 5+, django-ckeditor cũ có thể gặp warning. 
# Nên chuyển sang bản fork mới hoặc cấu hình kỹ. 
# Tạm thời dùng bản này, nếu lỗi thì đổi sang 'django-ckeditor-5'
django-ckeditor